requirements.txt
absl-py==0.12.0
astunparse==1.6.3
attrs==21.2.0
cached-property==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cloudpickle==1.6.0
colorama==0.4.4
contextlib2==21.6.0
cycler==0.10.0
decorator==5.0.9
dill==0.3.4
dm-tree==0.1.6
edward2==0.0.2
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-api-core==1.30.0
google-auth==1.32.1
google-auth-oauthlib==0.4.4
google-cloud-core==1.7.1
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
h5py==3.1.0
idna==2.10
importlib-metadata==4.6.1
importlib-resources==5.2.0
jax==0.2.16
joblib==1.0.1
keras-nightly==2.7.0.dev2021070900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
libclang==11.1.0
Markdown==3.3.4
matplotlib==3.4.2
ml-collections==0.1.0
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.0
Pillow==8.3.1
promise==2.3
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
robustness-metrics @ git+https://github.com/google-research/robustness_metrics.git@c08df4a86fee978d38ef6eba8a41b9e483000c5b
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
six==1.15.0
tabulate==0.8.9
tb-nightly==2.6.0a20210708
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-datasets==4.3.0
tensorflow-hub==0.12.0
tensorflow-metadata==1.1.0
termcolor==1.1.0
tf-estimator-nightly==2.7.0.dev2021070901
tf-nightly==2.7.0.dev20210708
tfa-nightly==0.14.0.dev20210707103333
tfp-nightly==0.14.0.dev20210630
threadpoolctl==2.1.0
tqdm==4.61.2
typeguard==2.12.1
typing-extensions==3.7.4.3
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.5.0
